refactor(infra): split 917-line setup.ts into SRP modules
core/infrastructure/setup.ts mixed seven distinct responsibilities
(AI-CLI install, Gemini config, Antigravity skill, Codex skill,
statusline, version migration, results display). Split into focused
modules under core/infrastructure/setup/ and reduced the original
file to a 201-line orchestrator.

Modules created:
  - setup/ai-cli-install.ts   — npm-installs the chosen AI CLI
  - setup/gemini.ts           — Gemini router cleanup + GEMINI.md merge
  - setup/antigravity.ts      — Antigravity SKILL.md install
  - setup/codex-skill.ts      — Codex SKILL.md install + verifyRouterReady
  - setup/statusline.ts       — modular Claude Code statusline
  - setup/version-migration.ts — bump cliVersion on every project
  - setup/results-display.ts  — per-provider summary block

Importers updated to point at the new modules directly (no barrel,
per project rule against re-exports):
  - services/sync-service.ts          → setup/codex-skill
  - services/doctor-service.ts        → setup/codex-skill
  - commands/setup/wizard.ts          → setup/codex-skill (dyn-imports)
  - commands/setup/install-status-line.ts → setup/statusline

Behavior unchanged:
  - typecheck clean
  - biome clean
  - 1029 tests pass (3483 expects)
  - bun run build succeeds
