+++ core/infrastructure/setup.ts
@@ -1,57 +1,46 @@
 /**
- * Setup Module - Core installation logic
+ * Setup orchestrator. Drives the per-provider installation pipeline
+ * (Claude Code, Gemini CLI, plus Antigravity & Codex skill installs)
+ * and finishes by migrating known projects to the current CLI version.
  *
- * Executes ALL setup needed for prjct-cli:
- * 1. Detect AI provider (Claude Code or Gemini CLI)
- * 2. Install CLI if missing
- * 3. Sync commands to provider's commands directory
- * 4. Install global config (CLAUDE.md or GEMINI.md)
- * 5. Save version in editors-config
+ * Implementation details live in `./setup/`:
+ *   - ai-cli-install.ts   — npm-installs the chosen AI CLI when missing
+ *   - gemini.ts           — Gemini router cleanup + GEMINI.md merge
+ *   - antigravity.ts      — Google Antigravity SKILL.md install
+ *   - codex-skill.ts      — OpenAI Codex SKILL.md install + verification
+ *   - statusline.ts       — modular Claude Code statusline install
+ *   - version-migration.ts — bump cliVersion on every known project
+ *   - results-display.ts  — per-provider summary block
  *
- * Supports multiple AI CLI agents:
- * - Claude Code: ~/.claude/commands/p.md, CLAUDE.md
- * - Gemini CLI: ~/.gemini/commands/p.toml, GEMINI.md
- *
- * This module is called from:
- * - core/index.js (on first CLI use)
- * - scripts/postinstall.js (if npm scripts are enabled)
+ * Exposed:
+ *   - `run()` — full pipeline. Auto-executes when this file is invoked
+ *     directly (`bun core/infrastructure/setup.ts`).
+ *   - `installCodexSkill`, `verifyCodexPRouterReady` — re-exported via
+ *     direct imports from `./setup/codex-skill` (no barrel here; per
+ *     project rule, callers import the module they actually need).
+ *   - `installStatusLine` — same: import from `./setup/statusline`.
  */
 
-import { execFileSync } from 'node:child_process'
-import fs from 'node:fs/promises'
-import os from 'node:os'
-import path from 'node:path'
 import chalk from 'chalk'
-import { getTemplateContent } from '../agentic/template-loader'
 import context7Service from '../services/context7-service'
-import { dependencyValidator } from '../services/dependency-validator'
-import { prjctDb } from '../storage/database'
-import { getErrorMessage, isNotFoundError } from '../types/fs'
-import type { AIProviderConfig, AIProviderName } from '../types/provider'
-import { getTimeout } from '../utils/constants'
-import { fileExists, readJson, writeJson } from '../utils/file-helper'
-import { sha256 } from '../utils/hash'
-import log from '../utils/logger'
-import { PACKAGE_ROOT, VERSION } from '../utils/version'
+import type { AIProviderName } from '../types/provider'
+import { VERSION } from '../utils/version'
 import {
   detectAllProviders,
   detectAntigravity,
   detectCodex,
-  detectProvider,
   Providers,
   selectProvider,
 } from './ai-provider'
 import installer from './command-installer'
 import editorsConfig from './editors-config'
-import { mergeWithMarkers } from './ide-project-installer'
-
-interface ProviderSetupResult {
-  provider: AIProviderName
-  cliInstalled: boolean
-  commandsAdded: number
-  commandsUpdated: number
-  configAction: string | null
-}
+import { installAICLI } from './setup/ai-cli-install'
+import { installAntigravitySkill } from './setup/antigravity'
+import { installCodexSkill, verifyCodexPRouterReady } from './setup/codex-skill'
+import { installGeminiGlobalConfig, installGeminiRouter } from './setup/gemini'
+import { type ProviderSetupResult, showResults } from './setup/results-display'
+import { installStatusLine } from './setup/statusline'
+import { migrateProjectsCliVersion } from './setup/version-migration'
 
 interface SetupResults {
   provider: AIProviderName // Primary provider
@@ -63,83 +52,12 @@ interface SetupResults {
 }
 
 /**
- * Check if an AI CLI is installed
- */
-async function _hasAICLI(provider: AIProviderConfig): Promise<boolean> {
-  const detection = await detectProvider(provider.name)
-  return detection.installed
-}
-
-/**
- * Install AI CLI for the specified provider
- * PRJ-114: Enhanced with graceful degradation and alternative install suggestions
- */
-async function installAICLI(provider: AIProviderConfig): Promise<boolean> {
-  const packageName =
-    provider.name === 'claude' ? '@anthropic-ai/claude-code' : '@google/gemini-cli'
-
-  // PRJ-114: Check npm availability first
-  if (!dependencyValidator.isAvailable('npm')) {
-    console.log(`${chalk.yellow('⚠️  npm is not available')}`)
-    console.log('')
-    console.log(`${chalk.dim(`Install ${provider.displayName} using one of:`)}`)
-    console.log(chalk.dim('  • Install Node.js: https://nodejs.org'))
-    console.log(
-      chalk.dim(
-        `  • Use Homebrew: brew install ${provider.name === 'claude' ? 'claude' : 'gemini'}`
-      )
-    )
-    console.log(chalk.dim(`  • Use npx directly: npx ${packageName}`))
-    console.log('')
-    return false
-  }
-
-  try {
-    console.log(chalk.yellow(`📦 ${provider.displayName} not found. Installing...`))
-    console.log('')
-    // PRJ-111: Add timeout to npm install (default: 2 minutes, configurable via PRJCT_TIMEOUT_NPM_INSTALL)
-    execFileSync('npm', ['install', '-g', packageName], {
-      stdio: 'inherit',
-      timeout: getTimeout('NPM_INSTALL'),
-    })
-    console.log('')
-    console.log(`${chalk.green('✓')} ${provider.displayName} installed successfully`)
-    console.log('')
-    return true
-  } catch (error) {
-    const err = error as Error & { killed?: boolean; signal?: string }
-    const isTimeout = err.killed && err.signal === 'SIGTERM'
-
-    if (isTimeout) {
-      console.log(chalk.yellow(`⚠️  Installation timed out for ${provider.displayName}`))
-      console.log('')
-      console.log(chalk.dim('The npm install took too long. Try:'))
-      console.log(chalk.dim('  • Set PRJCT_TIMEOUT_NPM_INSTALL=300000 for 5 minutes'))
-      console.log(chalk.dim(`  • Run manually: npm install -g ${packageName}`))
-    } else {
-      console.log(chalk.yellow(`⚠️  Failed to install ${provider.displayName}: ${err.message}`))
-    }
-    console.log('')
-    console.log(chalk.dim('Alternative installation methods:'))
-    console.log(chalk.dim(`  • npm:  npm install -g ${packageName}`))
-    console.log(chalk.dim(`  • yarn: yarn global add ${packageName}`))
-    console.log(chalk.dim(`  • pnpm: pnpm add -g ${packageName}`))
-    console.log(
-      chalk.dim(`  • brew: brew install ${provider.name === 'claude' ? 'claude' : 'gemini'}`)
-    )
-    console.log('')
-    return false
-  }
-}
-
-/**
- * Main setup function - installs for ALL detected providers
+ * Main setup function — installs for ALL detected providers.
  */
 async function run(): Promise<SetupResults> {
   // Step 0: Detect all available providers
   const detection = await detectAllProviders()
   const selection = await selectProvider()
-  const _primaryProvider = Providers[selection.provider]
 
   const results: SetupResults = {
     provider: selection.provider,
@@ -216,7 +134,7 @@ async function run(): Promise<SetupResults> {
         await context7Service.ensureReady()
       }
     } else if (providerName === 'gemini') {
-      // Gemini provider - install router and global config
+      // Gemini provider — install router and global config
       const geminiInstalled = await installGeminiRouter()
       if (geminiInstalled) {
         providerResult.commandsAdded = 1
@@ -272,642 +190,8 @@ async function run(): Promise<SetupResults> {
   return results
 }
 
-/**
- * Cleanup legacy Gemini router (p.toml) if it exists.
- * Router is deprecated — skills handle workflows natively.
- */
-async function installGeminiRouter(): Promise<boolean> {
-  try {
-    const geminiCommandsDir = path.join(os.homedir(), '.gemini', 'commands')
-    const routerDest = path.join(geminiCommandsDir, 'p.toml')
-
-    // Clean up legacy router if it exists
-    try {
-      await fs.unlink(routerDest)
-      return true
-    } catch (error) {
-      if ((error as NodeJS.ErrnoException).code === 'ENOENT') {
-        return false // Already gone
-      }
-      throw error
-    }
-  } catch (error) {
-    log.warn(`Gemini router cleanup warning: ${getErrorMessage(error)}`)
-    return false
-  }
-}
-
-/**
- * Install or update global GEMINI.md configuration
- */
-async function installGeminiGlobalConfig(): Promise<{ success: boolean; action: string | null }> {
-  try {
-    const geminiDir = path.join(os.homedir(), '.gemini')
-    const globalConfigPath = path.join(geminiDir, 'GEMINI.md')
-    // Ensure ~/.gemini directory exists
-    await fs.mkdir(geminiDir, { recursive: true })
-
-    // Read template content - try bundle first
-    let templateContent = getTemplateContent('global/GEMINI.md')
-    if (!templateContent) {
-      const templatePath = path.join(PACKAGE_ROOT, 'templates', 'global', 'GEMINI.md')
-      templateContent = await fs.readFile(templatePath, 'utf-8')
-    }
-
-    // Check if global config already exists
-    let existingContent = ''
-    let configExists = false
-
-    try {
-      existingContent = await fs.readFile(globalConfigPath, 'utf-8')
-      configExists = true
-    } catch (error) {
-      if (isNotFoundError(error)) {
-        configExists = false
-      } else {
-        throw error
-      }
-    }
-
-    const startMarker = '<!-- prjct:start - DO NOT REMOVE THIS MARKER -->'
-    const endMarker = '<!-- prjct:end - DO NOT REMOVE THIS MARKER -->'
-
-    const merged = mergeWithMarkers(
-      configExists ? existingContent : '',
-      templateContent,
-      startMarker,
-      endMarker
-    )
-
-    await fs.writeFile(globalConfigPath, merged.content, 'utf-8')
-    return { success: true, action: merged.action }
-  } catch (error) {
-    log.warn(`Gemini config warning: ${getErrorMessage(error)}`)
-    return { success: false, action: null }
-  }
-}
-
-// =============================================================================
-// Antigravity Installation (Skills-based)
-// =============================================================================
-
-/**
- * Install prjct as a skill for Google Antigravity
- *
- * Antigravity uses SKILL.md files in ~/.gemini/antigravity/skills/
- * This is the recommended integration method (not MCP).
- */
-async function installAntigravitySkill(): Promise<{
-  success: boolean
-  action: string | null
-}> {
-  try {
-    const antigravitySkillsDir = path.join(os.homedir(), '.gemini', 'antigravity', 'skills')
-    const prjctSkillDir = path.join(antigravitySkillsDir, 'prjct')
-    const skillMdPath = path.join(prjctSkillDir, 'SKILL.md')
-    // Ensure skills directory exists
-    await fs.mkdir(prjctSkillDir, { recursive: true })
-
-    // Check if SKILL.md already exists
-    const skillExists = await fileExists(skillMdPath)
-
-    // Read template content - try bundle first
-    let templateContent = getTemplateContent('antigravity/SKILL.md')
-    if (!templateContent) {
-      const templatePath = path.join(PACKAGE_ROOT, 'templates', 'antigravity', 'SKILL.md')
-      if (!(await fileExists(templatePath))) {
-        log.warn('Antigravity SKILL.md template not found')
-        return { success: false, action: null }
-      }
-      templateContent = await fs.readFile(templatePath, 'utf-8')
-    }
-
-    // Write SKILL.md
-    await fs.writeFile(skillMdPath, templateContent, 'utf-8')
-
-    return { success: true, action: skillExists ? 'updated' : 'created' }
-  } catch (error) {
-    log.warn(`Antigravity skill warning: ${getErrorMessage(error)}`)
-    return { success: false, action: null }
-  }
-}
-
-/**
- * Check if Antigravity skill needs installation or update
- */
-async function _needsAntigravityInstallation(): Promise<boolean> {
-  const detection = await detectAntigravity()
-  return detection.installed && !detection.skillInstalled
-}
-
-// =============================================================================
-// Codex Installation (Skills-based)
-// =============================================================================
-
-const CODEX_SKILL_META_MARKER = 'prjct-codex-router'
-
-import type { CodexPRouterStatus } from '../types/infrastructure.js'
-
-function getCodexSkillPath(): string {
-  return path.join(os.homedir(), '.codex', 'skills', 'prjct', 'SKILL.md')
-}
-
-function getCodexSkillMetadata(templateHash: string): string {
-  return `<!-- ${CODEX_SKILL_META_MARKER}: ${JSON.stringify({
-    version: VERSION,
-    templateHash,
-  })} -->`
-}
-
-function parseCodexSkillMetadata(
-  content: string
-): { version?: string; templateHash?: string } | null {
-  const match = content.match(
-    new RegExp(`<!--\\s*${CODEX_SKILL_META_MARKER}:\\s*(\\{[\\s\\S]*?\\})\\s*-->`)
-  )
-  if (!match) return null
-  try {
-    return JSON.parse(match[1]) as { version?: string; templateHash?: string }
-  } catch {
-    return null
-  }
-}
-
-function hashContent(content: string): string {
-  return sha256(content)
-}
-
-async function loadCodexSkillTemplate(): Promise<string | null> {
-  const bundled = getTemplateContent('codex/SKILL.md')
-  if (bundled) return bundled
-
-  const templatePath = path.join(PACKAGE_ROOT, 'templates', 'codex', 'SKILL.md')
-  if (!(await fileExists(templatePath))) {
-    return null
-  }
-
-  return fs.readFile(templatePath, 'utf-8')
-}
-
-function buildCodexSkillContent(templateContent: string): {
-  content: string
-  templateHash: string
-} {
-  const normalized = templateContent.trimEnd()
-  const templateHash = hashContent(normalized)
-  const metadata = getCodexSkillMetadata(templateHash)
-  return {
-    content: `${normalized}\n\n${metadata}\n`,
-    templateHash,
-  }
-}
-
-/**
- * Install prjct as a skill for OpenAI Codex
- *
- * Codex uses SKILL.md files in ~/.codex/skills/
- * Following the same pattern as Antigravity.
- */
-export async function installCodexSkill(): Promise<{
-  success: boolean
-  action: string | null
-}> {
-  try {
-    const skillMdPath = getCodexSkillPath()
-    const prjctSkillDir = path.dirname(skillMdPath)
-    // Ensure skills directory exists
-    await fs.mkdir(prjctSkillDir, { recursive: true })
-
-    // Check if SKILL.md already exists
-    const skillExists = await fileExists(skillMdPath)
-
-    // Read template content - try bundle first
-    const templateContent = await loadCodexSkillTemplate()
-    if (!templateContent) {
-      log.warn('Codex SKILL.md template not found')
-      return { success: false, action: null }
-    }
-
-    const built = buildCodexSkillContent(templateContent)
-
-    if (skillExists) {
-      const existing = await fs.readFile(skillMdPath, 'utf-8').catch(() => '')
-      if (existing === built.content) {
-        return { success: true, action: 'unchanged' }
-      }
-    }
-
-    // Write SKILL.md
-    await fs.writeFile(skillMdPath, built.content, 'utf-8')
-
-    return { success: true, action: skillExists ? 'updated' : 'created' }
-  } catch (error) {
-    log.warn(`Codex skill warning: ${getErrorMessage(error)}`)
-    return { success: false, action: null }
-  }
-}
-
-export async function verifyCodexPRouterReady(
-  options: { autoRepair?: boolean } = {}
-): Promise<CodexPRouterStatus> {
-  const skillPath = getCodexSkillPath()
-  const codexDetection = await detectCodex()
-  if (!codexDetection.installed) {
-    return {
-      installed: false,
-      verified: true,
-      skillPath,
-      message: 'Codex not detected',
-    }
-  }
-
-  const templateContent = await loadCodexSkillTemplate()
-  if (!templateContent) {
-    return {
-      installed: true,
-      verified: false,
-      skillPath,
-      message: 'Codex SKILL.md template missing from prjct installation',
-      fix: ['Reinstall prjct-cli package', 'Run `prjct setup`'],
-    }
-  }
-
-  const expected = buildCodexSkillContent(templateContent)
-
-  const maybeRepair = async (): Promise<boolean> => {
-    if (!options.autoRepair) return false
-    const result = await installCodexSkill()
-    return result.success
-  }
-
-  let skillContent = ''
-  if (!(await fileExists(skillPath))) {
-    if (!(await maybeRepair())) {
-      return {
-        installed: true,
-        verified: false,
-        skillPath,
-        templateHash: expected.templateHash,
-        message: 'Codex skill missing at ~/.codex/skills/prjct/SKILL.md',
-        fix: ['Run `prjct start` to install Codex skill'],
-      }
-    }
-  }
-
-  skillContent = await fs.readFile(skillPath, 'utf-8').catch(() => '')
-  let metadata = parseCodexSkillMetadata(skillContent)
-  const metadataMatches =
-    metadata?.version === VERSION && metadata?.templateHash === expected.templateHash
-
-  if (!metadataMatches) {
-    if (!(await maybeRepair())) {
-      return {
-        installed: true,
-        verified: false,
-        skillPath,
-        templateHash: expected.templateHash,
-        message: 'Codex skill metadata mismatch (outdated router)',
-        fix: ['Run `prjct start` or `prjct setup` to refresh Codex skill'],
-      }
-    }
-
-    skillContent = await fs.readFile(skillPath, 'utf-8').catch(() => '')
-    metadata = parseCodexSkillMetadata(skillContent)
-    const repaired =
-      metadata?.version === VERSION && metadata?.templateHash === expected.templateHash
-    if (!repaired) {
-      return {
-        installed: true,
-        verified: false,
-        skillPath,
-        templateHash: expected.templateHash,
-        message: 'Codex skill could not be repaired automatically',
-        fix: ['Delete ~/.codex/skills/prjct/SKILL.md', 'Run `prjct setup`'],
-      }
-    }
-  }
-
-  return {
-    installed: true,
-    verified: true,
-    skillPath,
-    templateHash: expected.templateHash,
-    message: 'Codex p. router ready',
-  }
-}
-
-/**
- * Migrate existing projects to add cliVersion field
- * This clears the status line warning after npm update
- */
-async function migrateProjectsCliVersion(): Promise<void> {
-  try {
-    const projectsDir = path.join(os.homedir(), '.prjct-cli', 'projects')
-
-    if (!(await fileExists(projectsDir))) {
-      return
-    }
-
-    const projectDirs = (await fs.readdir(projectsDir, { withFileTypes: true }))
-      .filter((dirent) => dirent.isDirectory())
-      .map((dirent) => dirent.name)
-
-    let migrated = 0
-
-    for (const projectId of projectDirs) {
-      try {
-        const project = prjctDb.getDoc<Record<string, unknown>>(projectId, 'project')
-        if (!project) {
-          continue
-        }
-
-        // Only update if cliVersion is missing or different
-        if (project.cliVersion !== VERSION) {
-          project.cliVersion = VERSION
-          prjctDb.setDoc(projectId, 'project', project)
-          migrated++
-        }
-      } catch {
-        // Skip projects with database issues
-      }
-    }
-
-    if (migrated > 0) {
-      console.log(`   ${chalk.green('✓')} Updated ${migrated} project(s) to v${VERSION}`)
-    }
-  } catch (error) {
-    // Silently fail if projects directory doesn't exist
-    if (!isNotFoundError(error)) {
-      // Log unexpected errors but don't crash - migration is optional
-      log.warn(`Migration warning: ${getErrorMessage(error)}`)
-    }
-  }
-}
-
-/**
- * Ensure settings.json has statusLine configured
- */
-async function ensureStatusLineSettings(
-  settingsPath: string,
-  statusLinePath: string
-): Promise<void> {
-  let settings: Record<string, unknown> = {}
-  if (await fileExists(settingsPath)) {
-    try {
-      settings = (await readJson<Record<string, unknown>>(settingsPath)) ?? {}
-    } catch (error) {
-      // Invalid JSON, start fresh - but propagate unexpected errors
-      if (!(error instanceof SyntaxError)) {
-        throw error
-      }
-    }
-  }
-  settings.statusLine = { type: 'command', command: statusLinePath }
-  await writeJson(settingsPath, settings)
-}
-
-/**
- * Install status line script with version check
- * Copies modular statusline from assets/ to ~/.prjct-cli/statusline/
- * Includes: statusline.sh, lib/, components/, themes/, config.json
- * Creates symlink at ~/.claude/prjct-statusline.sh
- * Updates CLI_VERSION in the script
- */
-export async function installStatusLine(): Promise<void> {
-  try {
-    const claudeDir = path.join(os.homedir(), '.claude')
-    const settingsPath = path.join(claudeDir, 'settings.json')
-    const claudeStatusLinePath = path.join(claudeDir, 'prjct-statusline.sh')
-
-    // Target location for the actual script
-    const prjctStatusLineDir = path.join(os.homedir(), '.prjct-cli', 'statusline')
-    const prjctStatusLinePath = path.join(prjctStatusLineDir, 'statusline.sh')
-    const prjctThemesDir = path.join(prjctStatusLineDir, 'themes')
-    const prjctLibDir = path.join(prjctStatusLineDir, 'lib')
-    const prjctComponentsDir = path.join(prjctStatusLineDir, 'components')
-    const prjctConfigPath = path.join(prjctStatusLineDir, 'config.json')
-
-    // Source assets (from the package)
-    const assetsDir = path.join(PACKAGE_ROOT, 'assets', 'statusline')
-    const sourceScript = path.join(assetsDir, 'statusline.sh')
-    const sourceThemeDir = path.join(assetsDir, 'themes')
-    const sourceLibDir = path.join(assetsDir, 'lib')
-    const sourceComponentsDir = path.join(assetsDir, 'components')
-    const sourceConfigPath = path.join(assetsDir, 'default-config.json')
-
-    // Ensure directories exist
-    if (!(await fileExists(claudeDir))) {
-      await fs.mkdir(claudeDir, { recursive: true })
-    }
-    if (!(await fileExists(prjctStatusLineDir))) {
-      await fs.mkdir(prjctStatusLineDir, { recursive: true })
-    }
-    if (!(await fileExists(prjctThemesDir))) {
-      await fs.mkdir(prjctThemesDir, { recursive: true })
-    }
-    if (!(await fileExists(prjctLibDir))) {
-      await fs.mkdir(prjctLibDir, { recursive: true })
-    }
-    if (!(await fileExists(prjctComponentsDir))) {
-      await fs.mkdir(prjctComponentsDir, { recursive: true })
-    }
-
-    // Check if statusline already exists
-    if (await fileExists(prjctStatusLinePath)) {
-      const existingContent = await fs.readFile(prjctStatusLinePath, 'utf8')
-
-      if (existingContent.includes('CLI_VERSION=')) {
-        // Has CLI_VERSION - update if needed
-        const versionMatch = existingContent.match(/CLI_VERSION="([^"]*)"/)
-
-        if (versionMatch && versionMatch[1] !== VERSION) {
-          // Update CLI_VERSION in-place
-          const updatedContent = existingContent.replace(
-            /CLI_VERSION="[^"]*"/,
-            `CLI_VERSION="${VERSION}"`
-          )
-          await fs.writeFile(prjctStatusLinePath, updatedContent, { mode: 0o755 })
-        }
-
-        // Ensure modular structure is installed (upgrade path)
-        await installStatusLineModules(sourceLibDir, prjctLibDir)
-        await installStatusLineModules(sourceComponentsDir, prjctComponentsDir)
-
-        // Ensure symlink and settings
-        await ensureStatusLineSymlink(claudeStatusLinePath, prjctStatusLinePath)
-        await ensureStatusLineSettings(settingsPath, claudeStatusLinePath)
-        return
-      }
-      // else: Script exists WITHOUT CLI_VERSION - fall through to replace with new version
-    }
-
-    // Install fresh from assets if source exists
-    if (await fileExists(sourceScript)) {
-      // Copy script and update version
-      let scriptContent = await fs.readFile(sourceScript, 'utf8')
-      scriptContent = scriptContent.replace(/CLI_VERSION="[^"]*"/, `CLI_VERSION="${VERSION}"`)
-      await fs.writeFile(prjctStatusLinePath, scriptContent, { mode: 0o755 })
-
-      // Copy lib/ modules
-      await installStatusLineModules(sourceLibDir, prjctLibDir)
-
-      // Copy components/
-      await installStatusLineModules(sourceComponentsDir, prjctComponentsDir)
-
-      // Copy themes
-      if (await fileExists(sourceThemeDir)) {
-        const themes = await fs.readdir(sourceThemeDir)
-        for (const theme of themes) {
-          const src = path.join(sourceThemeDir, theme)
-          const dest = path.join(prjctThemesDir, theme)
-          // Always update themes to get new icons/colors
-          await fs.copyFile(src, dest)
-        }
-      }
-
-      // Copy default config (only if not exists - preserve user customizations)
-      if (!(await fileExists(prjctConfigPath)) && (await fileExists(sourceConfigPath))) {
-        await fs.copyFile(sourceConfigPath, prjctConfigPath)
-      }
-    } else {
-      // Fallback: create simple script inline
-      const scriptContent = `#!/bin/bash
-# prjct Status Line for Claude Code
-CLI_VERSION="${VERSION}"
-input=$(cat)
-CWD=$(echo "$input" | jq -r '.workspace.current_dir // "~"' 2>/dev/null)
-CONFIG="$CWD/.prjct/prjct.config.json"
-if [ -f "$CONFIG" ]; then
-  PROJECT_ID=$(jq -r '.projectId // ""' "$CONFIG" 2>/dev/null)
-  if [ -n "$PROJECT_ID" ]; then
-    PROJECT_JSON="$HOME/.prjct-cli/projects/$PROJECT_ID/project.json"
-    if [ -f "$PROJECT_JSON" ]; then
-      PROJECT_VERSION=$(jq -r '.cliVersion // ""' "$PROJECT_JSON" 2>/dev/null)
-      if [ -z "$PROJECT_VERSION" ] || [ "$PROJECT_VERSION" != "$CLI_VERSION" ]; then
-        echo "prjct v$CLI_VERSION - run p. sync"
-        exit 0
-      fi
-    else
-      echo "prjct v$CLI_VERSION - run p. sync"
-      exit 0
-    fi
-    STATE="$HOME/.prjct-cli/projects/$PROJECT_ID/storage/state.json"
-    if [ -f "$STATE" ]; then
-      TASK=$(jq -r '.currentTask.description // ""' "$STATE" 2>/dev/null)
-      if [ -n "$TASK" ]; then
-        echo "$TASK"
-        exit 0
-      fi
-    fi
-  fi
-fi
-echo "prjct"
-`
-      await fs.writeFile(prjctStatusLinePath, scriptContent, { mode: 0o755 })
-    }
-
-    // Create symlink and configure settings
-    await ensureStatusLineSymlink(claudeStatusLinePath, prjctStatusLinePath)
-    await ensureStatusLineSettings(settingsPath, claudeStatusLinePath)
-  } catch (error) {
-    // Silently fail if directories don't exist
-    if (!isNotFoundError(error)) {
-      // Log unexpected errors but don't crash - status line is optional
-      log.warn(`Status line warning: ${getErrorMessage(error)}`)
-    }
-  }
-}
-
-/**
- * Install statusline modules (lib/ or components/)
- * Copies .sh files from source to destination, always overwriting for updates
- */
-async function installStatusLineModules(sourceDir: string, destDir: string): Promise<void> {
-  if (!(await fileExists(sourceDir))) {
-    return
-  }
-
-  const files = await fs.readdir(sourceDir)
-  for (const file of files) {
-    if (file.endsWith('.sh')) {
-      const src = path.join(sourceDir, file)
-      const dest = path.join(destDir, file)
-      await fs.copyFile(src, dest)
-      await fs.chmod(dest, 0o755)
-    }
-  }
-}
-
-/**
- * Ensure symlink from Claude config to prjct statusline
- */
-async function ensureStatusLineSymlink(linkPath: string, targetPath: string): Promise<void> {
-  try {
-    // Check if link already points to correct target
-    if (await fileExists(linkPath)) {
-      const stats = await fs.lstat(linkPath)
-      if (stats.isSymbolicLink()) {
-        const existingTarget = await fs.readlink(linkPath)
-        if (existingTarget === targetPath) {
-          return // Already correct
-        }
-      }
-      // Remove existing file/symlink
-      await fs.unlink(linkPath)
-    }
-    // Create symlink
-    await fs.symlink(targetPath, linkPath)
-  } catch (_error) {
-    // If symlink fails (e.g., Windows, permission issues), try copy instead
-    try {
-      if (await fileExists(targetPath)) {
-        await fs.copyFile(targetPath, linkPath)
-        await fs.chmod(linkPath, 0o755)
-      }
-    } catch (copyError) {
-      // Both symlink and copy failed - log if unexpected error
-      if (!isNotFoundError(copyError)) {
-        log.warn(`Symlink fallback warning: ${(copyError as Error).message}`)
-      }
-    }
-  }
-}
-
-/**
- * Show setup results for a single provider
- */
-function showResults(results: ProviderSetupResult, provider: AIProviderConfig): void {
-  console.log('')
-
-  if (results.cliInstalled) {
-    console.log(`   ${chalk.green('✓')} ${provider.displayName} CLI installed`)
-  } else {
-    console.log(`   ${chalk.green('✓')} ${provider.displayName} CLI found`)
-  }
-
-  const totalCommands = results.commandsAdded + results.commandsUpdated
-  if (totalCommands > 0) {
-    const parts: string[] = []
-    if (results.commandsAdded > 0) parts.push(`${results.commandsAdded} new`)
-    if (results.commandsUpdated > 0) parts.push(`${results.commandsUpdated} updated`)
-    console.log(`   ${chalk.green('✓')} Commands synced (${parts.join(', ')})`)
-  } else {
-    console.log(`   ${chalk.green('✓')} Commands up to date`)
-  }
-
-  if (results.configAction === 'created') {
-    console.log(`   ${chalk.green('✓')} Global config created (${provider.contextFile})`)
-  } else if (results.configAction === 'updated') {
-    console.log(`   ${chalk.green('✓')} Global config updated (${provider.contextFile})`)
-  } else if (results.configAction === 'appended') {
-    console.log(`   ${chalk.green('✓')} Global config merged (${provider.contextFile})`)
-  }
-
-  console.log('')
-}
-
-// Auto-execute when run directly (for bun/node CLI usage)
-// This enables: bun core/infrastructure/setup.ts
+// Auto-execute when run directly (for bun/node CLI usage).
+// This enables: `bun core/infrastructure/setup.ts`.
 const isDirectRun = process.argv[1]?.includes('setup.ts') || process.argv[1]?.includes('setup.js')
 if (isDirectRun) {
   run().catch((error) => {
