+++ core/services/doctor-service.ts
@@ -14,7 +14,7 @@ import path from 'node:path'
 import chalk from 'chalk'
 import configManager from '../infrastructure/config-manager'
 import pathManager from '../infrastructure/path-manager'
-import { verifyCodexPRouterReady } from '../infrastructure/setup'
+import { verifyCodexPRouterReady } from '../infrastructure/setup/codex-skill'
 import { stateStorage } from '../storage/state-storage'
 import type { CheckResult, DoctorResult } from '../types/services.js'
 import { fileExists } from '../utils/file-helper'
