+++ core/services/sync-service.ts
@@ -27,7 +27,7 @@ import { detectCodex } from '../infrastructure/ai-provider'
 import commandInstaller from '../infrastructure/command-installer'
 import configManager from '../infrastructure/config-manager'
 import pathManager from '../infrastructure/path-manager'
-import { verifyCodexPRouterReady } from '../infrastructure/setup'
+import { verifyCodexPRouterReady } from '../infrastructure/setup/codex-skill'
 import { analysisStorage } from '../storage/analysis-storage'
 import { archiveStorage } from '../storage/archive-storage'
 import { prjctDb } from '../storage/database'
